requirements.txt
torch~=1.7.0+cu101
torchvision~=0.8.0
future~=0.18.2
opencv-python~=4.5.3.56
matplotlib~=3.4.3
numpy~=1.20.3
visdom~=0.1.8.9
pillow~=8.3.2
tqdm~=4.62.3